requirements.txt
streamlit
pandas
joblib
scikit-learn
openai
python-dotenv
